Upgrade CSS generator to use background-size for >1 pixelRatio displays
The previous implementation attempted to use larger offsets for the
larger pixel density files. This also meant the actual elements
themselves needing to be larger which breaks normal layouts.

This allows the elements to stay the same size, but the backgrounds
to change size, thus preserving layout regardless of which density
file is loaded.

Tested on:
 iphone 2G (pixelDepth 1)
 iphone4 (pixelDepth 2)
 Photon Q (pixelDepth 1)
 Note 2 (pixelDepth 2):w